+++ mix/tasks/download_core_data.ex
@@ -1,29 +1,37 @@
 if File.exists?(Cldr.Config.download_data_dir()) do
-  defmodule Mix.Tasks.Cldr.Download.PluralRanges do
+  defmodule Mix.Tasks.Cldr.Download.CoreData do
     @moduledoc """
-    Downloads the pluralRanges.xml file from CLDR
+    Downloads the pluralRanges.xml and timezones.xml files from CLDR
     """
 
     use Mix.Task
     require Logger
 
-    @shortdoc "Downloads the pluralRanges.xml file from CLDR"
+    @shortdoc "Downloads the pluralRanges.xml and timezones.xml files from CLDR"
 
     @url 'https://unicode.org/Public/cldr/'
     @plural_range 'common/supplemental/pluralRanges.xml'
+    @timezones 'common/bcp47/timezone.xml'
     @core 'core.zip'
-    @output_file Path.join(Cldr.Config.download_data_dir(), "plural_ranges.xml")
+
+    @plural_ranges_output_file Path.join(Cldr.Config.download_data_dir(), "plural_ranges.xml")
+    @timezones_output_file Path.join(Cldr.Config.download_data_dir(), "timezones.xml")
 
     def run(_) do
       with {:ok, content} <- download(url()),
            :ok <- File.write(to_string(@core), content),
-           {:ok, [{@plural_range, plural_ranges}]} <-
-             :zip.unzip(@core, [:memory, {:file_list, [@plural_range]}]) do
-        File.write!(@output_file, plural_ranges)
-        Logger.info("Saved #{@plural_range} to #{@output_file}")
+           {:ok, [{@timezones, timezones}, {@plural_range, plural_ranges}]} <-
+             :zip.unzip(@core, [:memory, {:file_list, [@plural_range, @timezones]}]) do
+
+        File.write!(@plural_ranges_output_file, plural_ranges)
+        Logger.info("Saved #{@plural_range} to #{@plural_ranges_output_file}")
+
+        File.write!(@timezones_output_file, timezones)
+        Logger.info("Saved #{@timezones} to #{@timezones_output_file}")
+
         File.rm!(to_string(@core))
       else
-        {:error, reason} -> {:error, "Unable to write #{@output_file}: #{inspect reason}"}
+        {:error, reason} -> IO.inspect(reason)
       end
     end
 
@@ -31,12 +39,12 @@ if File.exists?(Cldr.Config.download_data_dir()) do
     def download(url) do
       case :httpc.request(url) do
         {:ok, {{_version, 200, 'OK'}, _headers, body}} ->
-          Logger.info("Downloaded pluralRanges.xml")
+          Logger.info("Downloaded CLDR core data")
           {:ok, body}
 
         {_, {{_version, code, message}, _headers, _body}} ->
           Logger.error(
-            "Failed to download pluralRanges.xml from #{url()}. " <>
+            "Failed to download CLDR core data from #{url()}. " <>
               "HTTP Error: (#{code}) #{inspect(message)}"
           )
 
@@ -45,7 +53,7 @@ if File.exists?(Cldr.Config.download_data_dir()) do
         {:error, {:failed_connect, [{_, {host, _port}}, {_, _, sys_message}]}} ->
           Logger.error(
             "Failed to connect to #{inspect(host)} to download " <>
-              "pluralRanges.xml. Reason: #{inspect(sys_message)}"
+              "CLDR core data. Reason: #{inspect(sys_message)}"
           )
 
           {:error, sys_message}