+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This is the changelog for Cldr v2.13.0 released on ______, 2020.  For older chan
 
 * Removes the runtime dependency on `Jason` since the RFC5646 parser is now inlined. Closes #99
 
+* Adds `Cldr.Timezone` to support timezone mapping for language tags using the `tz` key of the `u` extension to the [Unicode locale identifier](https://unicode.org/reports/tr35/#u_Extension)
+
 # Changelog for Cldr v2.12.1
 
 This is the changelog for Cldr v2.12.1 released on January 14th, 2020.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-cldr/cldr/tags)